Edit style.css for desired outcome of Lesson 6
Edited style.css to conform with desired outcome of The Odin Project Lesson 6. I want to come back to this and double check everything, but I beliee I have created the desired outcome. One thing I'm not sure of is if the buttons are meant to have black borders or not. I am leaving this for now since the README.md instructed me to not make edits to the styles, only the order or removing them.