Update 01-cascade-fix cascade_exercise_solution
The README has been updated in order to avoid saying that we provide
all possibles solutions, since this is not necessary.

The additional way to solve the last part of the exercise, which
was included to complete the solutions given, has been removed in order
to not bombard learners with too many options.